Added More Tests & Re-made Mayor
Since the Mayor role is turned-based, I converted the value from a role
to a boolean. This way when we implement our turn system we can use the
boolean mayor value to check if a player is the mayor or not. This will
help rotating the mayor role throughout turns.

Secondly, I added more tests regarding role changes and abilities of a
player created with a role (i.e new Player(“Captian”) vs. new Player())